+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Format: `## [vX.Y.Z] - YYYY-MM-DD` followed by categorized entries.
 
 ## [Unreleased]
 
+### Added
+- **Land Manager - Maintenance Gaurd ** When splinterlands is in maintenance mode disable land manager
+
 ---
 
 ## [v1.11.0] - 2026-06-08